# Faster and simpler pool.