--- a/export_test.go
+++ b/export_test.go
@@ -1,6 +1,10 @@
 package redis
 
-import "gopkg.in/redis.v3/internal/pool"
+import (
+	"time"
+
+	"gopkg.in/redis.v3/internal/pool"
+)
 
 func (c *baseClient) Pool() pool.Pooler {
 	return c.connPool
@@ -8,4 +12,8 @@ func (c *baseClient) Pool() pool.Pooler {
 
 func (c *PubSub) Pool() pool.Pooler {
 	return c.base.connPool
+}
+
+func SetReceiveMessageTimeout(d time.Duration) {
+	receiveMessageTimeout = d
 }