--- a/internal/pool/bench_test.go
+++ b/internal/pool/bench_test.go
@@ -2,7 +2,6 @@ package pool_test
 
 import (
 	"errors"
-	"net"
 	"testing"
 	"time"
 
@@ -10,22 +9,19 @@ import (
 )
 
 func benchmarkPoolGetPut(b *testing.B, poolSize int) {
-	dial := func() (net.Conn, error) {
-		return &net.TCPConn{}, nil
-	}
-	pool := pool.NewConnPool(dial, poolSize, time.Second, 0)
-	pool.DialLimiter = nil
+	connPool := pool.NewConnPool(dummyDialer, poolSize, time.Second, time.Hour, time.Hour)
+	connPool.DialLimiter = nil
 
 	b.ResetTimer()
 
 	b.RunParallel(func(pb *testing.PB) {
 		for pb.Next() {
-			conn, err := pool.Get()
+			cn, err := connPool.Get()
 			if err != nil {
-				b.Fatalf("no error expected on pool.Get but received: %s", err.Error())
+				b.Fatal(err)
 			}
-			if err = pool.Put(conn); err != nil {
-				b.Fatalf("no error expected on pool.Put but received: %s", err.Error())
+			if err = connPool.Put(cn); err != nil {
+				b.Fatal(err)
 			}
 		}
 	})
@@ -43,38 +39,34 @@ func BenchmarkPoolGetPut1000Conns(b *testing.B) {
 	benchmarkPoolGetPut(b, 1000)
 }
 
-func benchmarkPoolGetReplace(b *testing.B, poolSize int) {
-	dial := func() (net.Conn, error) {
-		return &net.TCPConn{}, nil
-	}
-	pool := pool.NewConnPool(dial, poolSize, time.Second, 0)
-	pool.DialLimiter = nil
-
+func benchmarkPoolGetRemove(b *testing.B, poolSize int) {
+	connPool := pool.NewConnPool(dummyDialer, poolSize, time.Second, time.Hour, time.Hour)
+	connPool.DialLimiter = nil
 	removeReason := errors.New("benchmark")
 
 	b.ResetTimer()
 
 	b.RunParallel(func(pb *testing.PB) {
 		for pb.Next() {
-			conn, err := pool.Get()
+			cn, err := connPool.Get()
 			if err != nil {
-				b.Fatalf("no error expected on pool.Get but received: %s", err.Error())
+				b.Fatal(err)
 			}
-			if err = pool.Replace(conn, removeReason); err != nil {
-				b.Fatalf("no error expected on pool.Remove but received: %s", err.Error())
+			if err := connPool.Remove(cn, removeReason); err != nil {
+				b.Fatal(err)
 			}
 		}
 	})
 }
 
-func BenchmarkPoolGetReplace10Conns(b *testing.B) {
-	benchmarkPoolGetReplace(b, 10)
+func BenchmarkPoolGetRemove10Conns(b *testing.B) {
+	benchmarkPoolGetRemove(b, 10)
 }
 
-func BenchmarkPoolGetReplace100Conns(b *testing.B) {
-	benchmarkPoolGetReplace(b, 100)
+func BenchmarkPoolGetRemove100Conns(b *testing.B) {
+	benchmarkPoolGetRemove(b, 100)
 }
 
-func BenchmarkPoolGetReplace1000Conns(b *testing.B) {
-	benchmarkPoolGetReplace(b, 1000)
+func BenchmarkPoolGetRemove1000Conns(b *testing.B) {
+	benchmarkPoolGetRemove(b, 1000)
 }