--- a/pool_test.go
+++ b/pool_test.go
@@ -106,7 +106,7 @@ var _ = Describe("pool", func() {
 
 		stats := pool.Stats()
 		Expect(stats.Requests).To(Equal(uint32(4)))
-		Expect(stats.Hits).To(Equal(uint32(3)))
+		Expect(stats.Hits).To(Equal(uint32(2)))
 		Expect(stats.Waits).To(Equal(uint32(0)))
 		Expect(stats.Timeouts).To(Equal(uint32(0)))
 	})