… internal/pool/conn_list.go
… internal/pool/conn_stack.go
… internal/pool/pool_test.go
… pubsub_test.go
… race_test.go
… redis_test.go