With recent pool changes we reintroduced `pool.Remove` which has flaw described in https://github.com/go-redis/redis/pull/89. So this PR fixes that flaw and also makes sure that `Dial` and `Conn.Close` is called without holding the lock.

Before

```
BenchmarkPoolGetPut10Conns-4      	 3000000	       439 ns/op	       0 B/op	       0 allocs/op
BenchmarkPoolGetPut100Conns-4     	 3000000	       409 ns/op	       0 B/op	       0 allocs/op
BenchmarkPoolGetPut1000Conns-4    	 5000000	       399 ns/op	       0 B/op	       0 allocs/op
BenchmarkPoolGetReplace10Conns-4  	  500000	      2925 ns/op	    8432 B/op	       7 allocs/op
BenchmarkPoolGetReplace100Conns-4 	  500000	      2918 ns/op	    8432 B/op	       7 allocs/op
BenchmarkPoolGetReplace1000Conns-4	  500000	      2938 ns/op	    8432 B/op	       7 allocs/op
```

After

```
BenchmarkPoolGetPut10Conns-4      	 5000000	       376 ns/op	       0 B/op	       0 allocs/op
BenchmarkPoolGetPut100Conns-4     	 5000000	       367 ns/op	       0 B/op	       0 allocs/op
BenchmarkPoolGetPut1000Conns-4    	 5000000	       377 ns/op	       0 B/op	       0 allocs/op
BenchmarkPoolGetReplace10Conns-4  	  300000	      4650 ns/op	    8419 B/op	       7 allocs/op
BenchmarkPoolGetReplace100Conns-4 	  300000	      4945 ns/op	    8419 B/op	       7 allocs/op
BenchmarkPoolGetReplace1000Conns-4	  300000	      4887 ns/op	    8419 B/op	       7 allocs/op
```

So it is slightly slower on remove, but I hope to make it faster in next PR.

PS Also I need to rewrite cluster reaper...

/cc @dim