@vmihailenco 2016-03-18
Cluster reaper fixed.